pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Chat interface for Mistral AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mistralai>=1.0.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6"
]
